Hide implementation details.
